Refactor LINQ generator to use a visitor pattern
